Use Vanilla .dll to check if Firewall is Disabled
We will still use the other functions there is no reason to change it for now.
