+++ GameLauncher/App/Classes/LauncherCore/InsiderKit/EnableInsider.cs
@@ -5,7 +5,7 @@ namespace GameLauncher.App.Classes.InsiderKit
         public static bool Insider = true;
 
         //Current month, day, year (2 digits), and letter! Ex: 12-15-20-A
-        public static string InsiderBuildNumber = "02-11-21-R";
+        public static string InsiderBuildNumber = "02-21-21-A";
 
         public static bool ShouldIBeAnInsider()
         {
